# docs/api/senado-openapi.json
{
	"openapi": "3.1.0",
	"info": {
		"title": "Dados Abertos Legislativos do Senado Federal e Congresso Nacional",
...
		"version": "4.0.3.54"
	},
	"servers": [
		{
			"url": "/",
			"description": "API de Dados Abertos Legislativos"
		}
	],
...
	"paths": {
...
		"/dadosabertos/materia/{sigla}/{numero}/{ano}": {
			"get": {
				"tags": [
					"Processo"
				],
				"summary": "(DEPRECATED) Detalhes da Matéria por Sigla, Número, Ano e Comissão",
				"description": "Substitua pelos serviços de listagem e detalhes de processo - [/dadosabertos/processo](#/Processo/processos) e [/dadosabertos/processo/{idProcesso}](#/Processo/detalhesProcesso)",
				"operationId": "detalhes",
				"parameters": [
					{
						"name": "sigla",
						"in": "path",
						"description": "Sigla do Tipo da Proposição",
						"required": true,
						"schema": {
							"type": "string"
						},
						"example": "REQ"
					},
					{
						"name": "numero",
						"in": "path",
						"description": "Número da Proposição",
						"required": true,
						"schema": {
							"type": "string"
						},
						"example": 1
					},
					{
						"name": "ano",
						"in": "path",
						"description": "Ano da Proposição",
						"required": true,
						"schema": {
							"type": "integer",
							"format": "int32"
						},
						"example": 2019
					},
					{
						"name": "comissao",
						"in": "query",
						"description": "Sigla da Comissão (somente para requerimentos de Comissão - sigla REQ)",
						"required": false,
						"schema": {
							"type": "string"
						},
						"example": "CCJ"
					},
					{
						"name": "v",
						"in": "query",
						"description": "Versão do serviço. Exemplificada a última disponível.",
						"required": false,
						"schema": {
							"type": "integer",
							"format": "int32",
							"default": 7
						},
						"example": 7
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/xml": {
								"schema": {
									"type": "string"
								}
							},
							"application/json": {
								"schema": {
									"type": "string"
								}
							},
							"text/csv": {
								"schema": {
									"type": "string"
								}
							},
							"*/*": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"deprecated": true
			}
		},
		"/dadosabertos/materia/{codigo}": {
			"get": {
				"tags": [
					"Processo"
				],
				"summary": "(DEPRECATED) Detalhes da Matéria por código",
				"description": "Substitua pelo serviço de detalhes do processo em [/dadosabertos/processo/{idProcesso}](#/Processo/detalhesProcesso)",
				"operationId": "detalhes_1",
				"parameters": [
					{
						"name": "codigo",
						"in": "path",
						"description": "Código da Matéria",
						"required": true,
						"schema": {
							"type": "integer",
							"format": "int64"
						},
						"example": 110428
					},
					{
						"name": "v",
						"in": "query",
						"description": "Versão do serviço. Exemplificada a última disponível.",
						"required": false,
						"schema": {
							"type": "integer",
							"format": "int32",
							"default": 7
						},
						"example": 7
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/xml": {
								"schema": {
									"type": "string"
								}
							},
							"application/json": {
								"schema": {
									"type": "string"
								}
							},
							"text/csv": {
								"schema": {
									"type": "string"
								}
							},
							"*/*": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"deprecated": true
			}
		},
...
	},
...
}